Drop LeveledLogger constructor inheritance
It works fine, but for some reason gcc stops the diagnostic "required
from here" trace at the inherited constructor, which makes it rather
difficult to figure out where a formatting error from caller code is
coming from.  If we avoid inheriting the constructor and just inline the
code gcc is happy to point at the proper call location again.