Support a list of source root prefixes
Some code bases have code in multiple places (e.g. src/ and include/)
but we have no way to strip them so would have to strip the project root
and have the `src/` or `include/` prefixed on everything.

This commit allows a list of prefixes and attempts to remove them all.
It also tweaks the way "../" and "/" get removed which ought to make it
slightly more robust against weird paths.