fmt/spdlog version bump and compatibility fixes
Our bundled spdlog doesn't work with fmt 10, and so if we picked up a
system fmt 10 then we'd allow either an incompatible spdlog, or use our
own (incompatible) spdlog submodule.

This fixes it by bumping our bundled versions to fmt 10 and spdlog 1.12
(which work together), and handles the increased minimum spdlog version
if we bundle or use a system fmt >= 10.