Fix crash at destruction with multiple static libs
If compiling oxen-logging with multiple instances statically (such as
libsession-util which includes a static libquic.a, itself with its own
oxen-logging.a), the `loggers` destructor would fire multiple times on
the same `loggers` instance, because of static linking rules.

Normally, when static linking, each instance would be in its own private
address space and so it would be two unrelated destructions on two
unrelated objects.  However the OXEN_LOGGING_EXPORT here forced them to
have the same address, but that doesn't defer the destruction, and so we
would get two destructions (and, also, two constructions) at the same
instance, which is bad.

This fixes it by hiding the unordered_map inside an exported function,
so that we get the same unique-instance behaviour, but there should only
ever be one instance because there should only be one function here that
gets called, and only one associated destruction because the static copy
in the other static lib never gets initialized.