Keep two parts when path stripping, except for src
Currently using only the basename in logs in a bit too restrictive: this
relaxes it to keep the basename and last part of the path, so that we
can have things like "server/omq.cpp" and "snode/swarm.cpp" instead of
just "omq.cpp" and "swarm.cpp".

We still keep only the filename, though, if the last part of the path
ends up being "src/" because that's useless, and is commonly found in
several projects (most notably libsession, libquic).